+++ test/dummy/app/models/binary_json_store_record.rb
@@ -1,3 +1,5 @@
 class BinaryJsonStoreRecord < ApplicationRecord
   include StructuredStore::Storable
+
+  structured_store :store
 end