+++ test/dummy/app/models/store_record.rb
@@ -1,6 +1,8 @@
-# This class demostrates using the StructuredStore to manage properties stored in a JSON field
+# This class demonstrates using the StructuredStore to manage properties stored in a JSON field
 class StoreRecord < ApplicationRecord
   include StructuredStore::Storable
+
+  structured_store :store
 
   # Returns a memoized instance of ChronicDateRangeConverter used for converting date ranges stored as a hash in the StructuredStore JSON field.
   # Other date range converters can be used.