+++ test/dummy/db/migrate/20250530114142_create_store_records.rb
@@ -1,8 +1,9 @@
 class CreateStoreRecords < ActiveRecord::Migration[7.2]
   def change
     create_table :store_records do |t|
-      t.references :structured_store_versioned_schema, null: false, foreign_key: true,
-                                                       index: { name: 'index_ss_records_on_versioned_schema_id' }
+      t.references :structured_store_store_versioned_schema,
+                   null: false,
+                   foreign_key: { to_table: :structured_store_versioned_schemas }
       t.json :store
 
       t.timestamps