+++ test/dummy/test/ref_resolvers/blank_ref_resolver_test.rb
@@ -14,7 +14,7 @@ module StructuredStore
           'properties' => {}
         }
 
-        store_record = StoreRecord.new(versioned_schema: VersionedSchema.new(json_schema: schema))
+        store_record = StoreRecord.new(store_versioned_schema: VersionedSchema.new(json_schema: schema))
 
         # Ensure the structure store attribute is not defined
         assert_not_respond_to store_record, :foo
@@ -33,7 +33,7 @@ module StructuredStore
           }
         }
 
-        store_record = StoreRecord.new(versioned_schema: VersionedSchema.new(json_schema: schema))
+        store_record = StoreRecord.new(store_versioned_schema: VersionedSchema.new(json_schema: schema))
 
         # Now the structured store attribute "foo" should be defined
         assert_nil store_record.foo
@@ -53,7 +53,7 @@ module StructuredStore
         }
 
         exception = assert_raises(RuntimeError) do
-          StoreRecord.new(versioned_schema: VersionedSchema.new(json_schema: schema))
+          StoreRecord.new(store_versioned_schema: VersionedSchema.new(json_schema: schema))
         end
 
         assert_equal 'Unsupported attribute type: "object" for property \'foo\'', exception.message